tests: Move some tests out of Other
