+++ tests/unit/debug_extensions.cpp
@@ -1,0 +1,321 @@
+/*
+ * Copyright (c) 2015-2023 The Khronos Group Inc.
+ * Copyright (c) 2015-2023 Valve Corporation
+ * Copyright (c) 2015-2023 LunarG, Inc.
+ * Copyright (c) 2015-2023 Google, Inc.
+ * Modifications Copyright (C) 2020-2021 Advanced Micro Devices, Inc. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ */
+
+#include "../framework/layer_validation_tests.h"
+
+TEST_F(NegativeDebugExtensions, DebugMarkerName) {
+    TEST_DESCRIPTION("Ensure debug marker object names are printed in debug report output");
+
+    AddRequiredExtensions(VK_EXT_DEBUG_REPORT_EXTENSION_NAME);
+    AddRequiredExtensions(VK_EXT_DEBUG_MARKER_EXTENSION_NAME);
+    RETURN_IF_SKIP(InitFramework());
+    RETURN_IF_SKIP(InitState());
+
+    if (IsPlatformMockICD()) {
+        GTEST_SKIP() << "Skipping object naming test with MockICD.";
+    }
+
+    VkBuffer buffer;
+    VkDeviceMemory memory_1, memory_2;
+    std::string memory_name = "memory_name";
+
+    VkBufferCreateInfo buffer_create_info = vku::InitStructHelper();
+    buffer_create_info.usage = VK_BUFFER_USAGE_UNIFORM_BUFFER_BIT;
+    buffer_create_info.size = 1;
+
+    vk::CreateBuffer(device(), &buffer_create_info, nullptr, &buffer);
+
+    VkMemoryRequirements memRequirements;
+    vk::GetBufferMemoryRequirements(device(), buffer, &memRequirements);
+
+    VkMemoryAllocateInfo memory_allocate_info = vku::InitStructHelper();
+    memory_allocate_info.allocationSize = memRequirements.size;
+    memory_allocate_info.memoryTypeIndex = 0;
+
+    vk::AllocateMemory(device(), &memory_allocate_info, nullptr, &memory_1);
+    vk::AllocateMemory(device(), &memory_allocate_info, nullptr, &memory_2);
+
+    VkDebugMarkerObjectNameInfoEXT name_info = vku::InitStructHelper();
+    name_info.object = (uint64_t)memory_2;
+    name_info.objectType = VK_DEBUG_REPORT_OBJECT_TYPE_DEVICE_MEMORY_EXT;
+    name_info.pObjectName = memory_name.c_str();
+    vk::DebugMarkerSetObjectNameEXT(device(), &name_info);
+
+    vk::BindBufferMemory(device(), buffer, memory_1, 0);
+
+    // Test core_validation layer
+    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, memory_name);
+    vk::BindBufferMemory(device(), buffer, memory_2, 0);
+    m_errorMonitor->VerifyFound();
+
+    vk::FreeMemory(device(), memory_1, nullptr);
+    memory_1 = VK_NULL_HANDLE;
+    vk::FreeMemory(device(), memory_2, nullptr);
+    memory_2 = VK_NULL_HANDLE;
+    vk::DestroyBuffer(device(), buffer, nullptr);
+    buffer = VK_NULL_HANDLE;
+
+    VkCommandBuffer commandBuffer;
+    std::string commandBuffer_name = "command_buffer_name";
+    VkCommandPoolCreateInfo pool_create_info = vku::InitStructHelper();
+    pool_create_info.queueFamilyIndex = m_device->graphics_queue_node_index_;
+    pool_create_info.flags = VK_COMMAND_POOL_CREATE_RESET_COMMAND_BUFFER_BIT;
+    vkt::CommandPool command_pool_1(*m_device, pool_create_info);
+    vkt::CommandPool command_pool_2(*m_device, pool_create_info);
+
+    VkCommandBufferAllocateInfo command_buffer_allocate_info = vku::InitStructHelper();
+    command_buffer_allocate_info.commandPool = command_pool_1.handle();
+    command_buffer_allocate_info.commandBufferCount = 1;
+    command_buffer_allocate_info.level = VK_COMMAND_BUFFER_LEVEL_PRIMARY;
+    vk::AllocateCommandBuffers(device(), &command_buffer_allocate_info, &commandBuffer);
+
+    name_info.object = (uint64_t)commandBuffer;
+    name_info.objectType = VK_DEBUG_REPORT_OBJECT_TYPE_COMMAND_BUFFER_EXT;
+    name_info.pObjectName = commandBuffer_name.c_str();
+    vk::DebugMarkerSetObjectNameEXT(device(), &name_info);
+
+    VkCommandBufferBeginInfo cb_begin_Info = vku::InitStructHelper();
+    cb_begin_Info.flags = VK_COMMAND_BUFFER_USAGE_ONE_TIME_SUBMIT_BIT;
+    vk::BeginCommandBuffer(commandBuffer, &cb_begin_Info);
+
+    const VkRect2D scissor = {{-1, 0}, {16, 16}};
+    const VkRect2D scissors[] = {scissor, scissor};
+
+    // Test parameter_validation layer
+    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, commandBuffer_name);
+    vk::CmdSetScissor(commandBuffer, 0, 1, scissors);
+    m_errorMonitor->VerifyFound();
+
+    // Test object_tracker layer
+    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, commandBuffer_name);
+    vk::FreeCommandBuffers(device(), command_pool_2.handle(), 1, &commandBuffer);
+    m_errorMonitor->VerifyFound();
+}
+
+TEST_F(NegativeDebugExtensions, DebugUtilsName) {
+    TEST_DESCRIPTION("Ensure debug utils object names are printed in debug messenger output");
+
+    AddRequiredExtensions(VK_EXT_DEBUG_UTILS_EXTENSION_NAME);
+
+    RETURN_IF_SKIP(InitFramework());
+    RETURN_IF_SKIP(InitState());
+
+    if (IsPlatformMockICD()) {
+        GTEST_SKIP() << "Skipping object naming test with MockICD.";
+    }
+
+    DebugUtilsLabelCheckData callback_data;
+    auto empty_callback = [](const VkDebugUtilsMessengerCallbackDataEXT *pCallbackData, DebugUtilsLabelCheckData *data) {
+        data->count++;
+    };
+    callback_data.count = 0;
+    callback_data.callback = empty_callback;
+
+    VkDebugUtilsMessengerCreateInfoEXT callback_create_info = vku::InitStructHelper();
+    callback_create_info.messageSeverity =
+        VK_DEBUG_UTILS_MESSAGE_SEVERITY_ERROR_BIT_EXT | VK_DEBUG_UTILS_MESSAGE_SEVERITY_WARNING_BIT_EXT;
+    callback_create_info.messageType = VK_DEBUG_UTILS_MESSAGE_TYPE_VALIDATION_BIT_EXT;
+    callback_create_info.pfnUserCallback = DebugUtilsCallback;
+    callback_create_info.pUserData = &callback_data;
+    VkDebugUtilsMessengerEXT my_messenger = VK_NULL_HANDLE;
+    vk::CreateDebugUtilsMessengerEXT(instance(), &callback_create_info, nullptr, &my_messenger);
+
+    VkBuffer buffer;
+    VkDeviceMemory memory_1, memory_2;
+    std::string memory_name = "memory_name";
+
+    VkBufferCreateInfo buffer_create_info = vku::InitStructHelper();
+    buffer_create_info.usage = VK_BUFFER_USAGE_UNIFORM_BUFFER_BIT;
+    buffer_create_info.size = 1;
+
+    vk::CreateBuffer(device(), &buffer_create_info, nullptr, &buffer);
+
+    VkMemoryRequirements memRequirements;
+    vk::GetBufferMemoryRequirements(device(), buffer, &memRequirements);
+
+    VkMemoryAllocateInfo memory_allocate_info = vku::InitStructHelper();
+    memory_allocate_info.allocationSize = memRequirements.size;
+    memory_allocate_info.memoryTypeIndex = 0;
+
+    vk::AllocateMemory(device(), &memory_allocate_info, nullptr, &memory_1);
+    vk::AllocateMemory(device(), &memory_allocate_info, nullptr, &memory_2);
+
+    VkDebugUtilsObjectNameInfoEXT name_info = vku::InitStructHelper();
+    name_info.objectType = VK_OBJECT_TYPE_DEVICE_MEMORY;
+    name_info.pObjectName = memory_name.c_str();
+
+    // Pass in bad handle make sure ObjectTracker catches it
+    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-VkDebugUtilsObjectNameInfoEXT-objectType-02590");
+    name_info.objectHandle = (uint64_t)0xcadecade;
+    vk::SetDebugUtilsObjectNameEXT(device(), &name_info);
+    m_errorMonitor->VerifyFound();
+
+    // Pass in null handle
+    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-vkSetDebugUtilsObjectNameEXT-pNameInfo-02588");
+    name_info.objectHandle = 0;
+    vk::SetDebugUtilsObjectNameEXT(device(), &name_info);
+    m_errorMonitor->VerifyFound();
+
+    // Pass in 'unknown' object type and see if parameter validation catches it
+    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-vkSetDebugUtilsObjectNameEXT-pNameInfo-02587");
+    name_info.objectHandle = (uint64_t)memory_2;
+    name_info.objectType = VK_OBJECT_TYPE_UNKNOWN;
+    vk::SetDebugUtilsObjectNameEXT(device(), &name_info);
+    m_errorMonitor->VerifyFound();
+
+    name_info.objectType = VK_OBJECT_TYPE_DEVICE_MEMORY;
+    vk::SetDebugUtilsObjectNameEXT(device(), &name_info);
+
+    vk::BindBufferMemory(device(), buffer, memory_1, 0);
+
+    // Test core_validation layer
+    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, memory_name);
+    vk::BindBufferMemory(device(), buffer, memory_2, 0);
+    m_errorMonitor->VerifyFound();
+
+    vk::FreeMemory(device(), memory_1, nullptr);
+    memory_1 = VK_NULL_HANDLE;
+    vk::FreeMemory(device(), memory_2, nullptr);
+    memory_2 = VK_NULL_HANDLE;
+    vk::DestroyBuffer(device(), buffer, nullptr);
+    buffer = VK_NULL_HANDLE;
+
+    VkCommandBuffer commandBuffer;
+    std::string commandBuffer_name = "command_buffer_name";
+    VkCommandPoolCreateInfo pool_create_info = vku::InitStructHelper();
+    pool_create_info.queueFamilyIndex = m_device->graphics_queue_node_index_;
+    pool_create_info.flags = VK_COMMAND_POOL_CREATE_RESET_COMMAND_BUFFER_BIT;
+    vkt::CommandPool command_pool_1(*m_device, pool_create_info);
+    vkt::CommandPool command_pool_2(*m_device, pool_create_info);
+
+    VkCommandBufferAllocateInfo command_buffer_allocate_info = vku::InitStructHelper();
+    command_buffer_allocate_info.commandPool = command_pool_1.handle();
+    command_buffer_allocate_info.commandBufferCount = 1;
+    command_buffer_allocate_info.level = VK_COMMAND_BUFFER_LEVEL_PRIMARY;
+    vk::AllocateCommandBuffers(device(), &command_buffer_allocate_info, &commandBuffer);
+
+    name_info.objectHandle = (uint64_t)commandBuffer;
+    name_info.objectType = VK_OBJECT_TYPE_COMMAND_BUFFER;
+    name_info.pObjectName = commandBuffer_name.c_str();
+    vk::SetDebugUtilsObjectNameEXT(device(), &name_info);
+
+    VkCommandBufferBeginInfo cb_begin_Info = vku::InitStructHelper();
+    cb_begin_Info.flags = VK_COMMAND_BUFFER_USAGE_ONE_TIME_SUBMIT_BIT;
+    vk::BeginCommandBuffer(commandBuffer, &cb_begin_Info);
+
+    const VkRect2D scissor = {{-1, 0}, {16, 16}};
+    const VkRect2D scissors[] = {scissor, scissor};
+
+    VkDebugUtilsLabelEXT command_label = vku::InitStructHelper();
+    command_label.pLabelName = "Command Label 0123";
+    command_label.color[0] = 0.;
+    command_label.color[1] = 1.;
+    command_label.color[2] = 2.;
+    command_label.color[3] = 3.0;
+    bool command_label_test = false;
+    auto command_label_callback = [command_label, &command_label_test](const VkDebugUtilsMessengerCallbackDataEXT *pCallbackData,
+                                                                       DebugUtilsLabelCheckData *data) {
+        data->count++;
+        command_label_test = false;
+        if (pCallbackData->cmdBufLabelCount == 1) {
+            command_label_test = pCallbackData->pCmdBufLabels[0] == command_label;
+        }
+    };
+    callback_data.callback = command_label_callback;
+
+    vk::CmdInsertDebugUtilsLabelEXT(commandBuffer, &command_label);
+    // Test parameter_validation layer
+    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, commandBuffer_name);
+    vk::CmdSetScissor(commandBuffer, 0, 1, scissors);
+    m_errorMonitor->VerifyFound();
+
+    // Check the label test
+    if (!command_label_test) {
+        ADD_FAILURE() << "Command label '" << command_label.pLabelName << "' not passed to callback.";
+    }
+
+    // Test object_tracker layer
+    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, commandBuffer_name);
+    vk::FreeCommandBuffers(device(), command_pool_2.handle(), 1, &commandBuffer);
+    m_errorMonitor->VerifyFound();
+
+    vk::DestroyDebugUtilsMessengerEXT(instance(), my_messenger, nullptr);
+}
+
+TEST_F(NegativeDebugExtensions, DebugUtilsParameterFlags) {
+    AddRequiredExtensions(VK_EXT_DEBUG_UTILS_EXTENSION_NAME);
+    RETURN_IF_SKIP(Init());
+
+    DebugUtilsLabelCheckData callback_data;
+    auto empty_callback = [](const VkDebugUtilsMessengerCallbackDataEXT *pCallbackData, DebugUtilsLabelCheckData *data) {
+        data->count++;
+    };
+    callback_data.count = 0;
+    callback_data.callback = empty_callback;
+
+    VkDebugUtilsMessengerCreateInfoEXT callback_create_info = vku::InitStructHelper();
+    callback_create_info.messageSeverity = 0;
+    callback_create_info.messageType = VK_DEBUG_UTILS_MESSAGE_TYPE_VALIDATION_BIT_EXT;
+    callback_create_info.pfnUserCallback = DebugUtilsCallback;
+    callback_create_info.pUserData = &callback_data;
+    VkDebugUtilsMessengerEXT my_messenger = VK_NULL_HANDLE;
+    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-VkDebugUtilsMessengerCreateInfoEXT-messageSeverity-requiredbitmask");
+    vk::CreateDebugUtilsMessengerEXT(instance(), &callback_create_info, nullptr, &my_messenger);
+    m_errorMonitor->VerifyFound();
+}
+
+struct LayerStatusCheckData {
+    std::function<void(const VkDebugUtilsMessengerCallbackDataEXT *pCallbackData, LayerStatusCheckData *)> callback;
+    ErrorMonitor *error_monitor;
+};
+
+TEST_F(NegativeDebugExtensions, LayerInfoMessages) {
+    TEST_DESCRIPTION("Ensure layer prints startup status messages.");
+
+    auto ici = GetInstanceCreateInfo();
+    LayerStatusCheckData callback_data;
+    auto local_callback = [](const VkDebugUtilsMessengerCallbackDataEXT *pCallbackData, LayerStatusCheckData *data) {
+        std::string message(pCallbackData->pMessage);
+        if ((data->error_monitor->GetMessageFlags() & kInformationBit) &&
+            (message.find("UNASSIGNED-khronos-validation-createinstance-status-message") == std::string::npos)) {
+            data->error_monitor->SetError("UNASSIGNED-Khronos-validation-createinstance-status-message-not-found");
+        } else if ((data->error_monitor->GetMessageFlags() & kPerformanceWarningBit) &&
+                   (message.find("UNASSIGNED-khronos-Validation-debug-build-warning-message") == std::string::npos)) {
+            data->error_monitor->SetError("UNASSIGNED-khronos-validation-createinstance-debug-warning-message-not-found");
+        }
+    };
+    callback_data.error_monitor = m_errorMonitor;
+    callback_data.callback = local_callback;
+
+    VkInstance local_instance;
+
+    VkDebugUtilsMessengerCreateInfoEXT callback_create_info = vku::InitStructHelper();
+    callback_create_info.messageSeverity = VK_DEBUG_UTILS_MESSAGE_SEVERITY_INFO_BIT_EXT;
+    callback_create_info.messageType = VK_DEBUG_UTILS_MESSAGE_TYPE_VALIDATION_BIT_EXT;
+    callback_create_info.pfnUserCallback = DebugUtilsCallback;
+    callback_create_info.pUserData = &callback_data;
+    ici.pNext = &callback_create_info;
+
+    // Create an instance, error if layer status INFO message not found
+    ASSERT_EQ(VK_SUCCESS, vk::CreateInstance(&ici, nullptr, &local_instance));
+    vk::DestroyInstance(local_instance, nullptr);
+
+#ifndef NDEBUG
+    // Create an instance, error if layer DEBUG_BUILD warning message not found
+    callback_create_info.messageSeverity = VK_DEBUG_UTILS_MESSAGE_SEVERITY_WARNING_BIT_EXT;
+    callback_create_info.messageType = VK_DEBUG_UTILS_MESSAGE_TYPE_PERFORMANCE_BIT_EXT;
+    ASSERT_EQ(VK_SUCCESS, vk::CreateInstance(&ici, nullptr, &local_instance));
+    vk::DestroyInstance(local_instance, nullptr);
+#endif
+}
