+++ tests/unit/others.cpp
@@ -471,51 +471,6 @@ TEST_F(VkLayerTest, VuidFilterInt) {
     TestRenderPassCreate(m_errorMonitor, *m_device, rpci, false, "VUID-VkInputAttachmentAspectReference-aspectMask-01964", nullptr);
 }
 
-struct LayerStatusCheckData {
-    std::function<void(const VkDebugUtilsMessengerCallbackDataEXT *pCallbackData, LayerStatusCheckData *)> callback;
-    ErrorMonitor *error_monitor;
-};
-
-TEST_F(VkLayerTest, LayerInfoMessages) {
-    TEST_DESCRIPTION("Ensure layer prints startup status messages.");
-
-    auto ici = GetInstanceCreateInfo();
-    LayerStatusCheckData callback_data;
-    auto local_callback = [](const VkDebugUtilsMessengerCallbackDataEXT *pCallbackData, LayerStatusCheckData *data) {
-        std::string message(pCallbackData->pMessage);
-        if ((data->error_monitor->GetMessageFlags() & kInformationBit) &&
-            (message.find("UNASSIGNED-khronos-validation-createinstance-status-message") == std::string::npos)) {
-            data->error_monitor->SetError("UNASSIGNED-Khronos-validation-createinstance-status-message-not-found");
-        } else if ((data->error_monitor->GetMessageFlags() & kPerformanceWarningBit) &&
-                   (message.find("UNASSIGNED-khronos-Validation-debug-build-warning-message") == std::string::npos)) {
-            data->error_monitor->SetError("UNASSIGNED-khronos-validation-createinstance-debug-warning-message-not-found");
-        }
-    };
-    callback_data.error_monitor = m_errorMonitor;
-    callback_data.callback = local_callback;
-
-    VkInstance local_instance;
-
-    VkDebugUtilsMessengerCreateInfoEXT callback_create_info = vku::InitStructHelper();
-    callback_create_info.messageSeverity = VK_DEBUG_UTILS_MESSAGE_SEVERITY_INFO_BIT_EXT;
-    callback_create_info.messageType = VK_DEBUG_UTILS_MESSAGE_TYPE_VALIDATION_BIT_EXT;
-    callback_create_info.pfnUserCallback = DebugUtilsCallback;
-    callback_create_info.pUserData = &callback_data;
-    ici.pNext = &callback_create_info;
-
-    // Create an instance, error if layer status INFO message not found
-    ASSERT_EQ(VK_SUCCESS, vk::CreateInstance(&ici, nullptr, &local_instance));
-    vk::DestroyInstance(local_instance, nullptr);
-
-#ifndef NDEBUG
-    // Create an instance, error if layer DEBUG_BUILD warning message not found
-    callback_create_info.messageSeverity = VK_DEBUG_UTILS_MESSAGE_SEVERITY_WARNING_BIT_EXT;
-    callback_create_info.messageType = VK_DEBUG_UTILS_MESSAGE_TYPE_PERFORMANCE_BIT_EXT;
-    ASSERT_EQ(VK_SUCCESS, vk::CreateInstance(&ici, nullptr, &local_instance));
-    vk::DestroyInstance(local_instance, nullptr);
-#endif
-}
-
 TEST_F(VkLayerTest, RequiredParameter) {
     TEST_DESCRIPTION("Specify VK_NULL_HANDLE, NULL, and 0 for required handle, pointer, array, and array count parameters");
 
@@ -791,267 +746,6 @@ TEST_F(VkLayerTest, ReservedParameter) {
     m_errorMonitor->VerifyFound();
 }
 
-TEST_F(VkLayerTest, DebugMarkerNameTest) {
-    TEST_DESCRIPTION("Ensure debug marker object names are printed in debug report output");
-
-    AddRequiredExtensions(VK_EXT_DEBUG_REPORT_EXTENSION_NAME);
-    AddRequiredExtensions(VK_EXT_DEBUG_MARKER_EXTENSION_NAME);
-    RETURN_IF_SKIP(InitFramework());
-    RETURN_IF_SKIP(InitState());
-
-    if (IsPlatformMockICD()) {
-        GTEST_SKIP() << "Skipping object naming test with MockICD.";
-    }
-
-    VkBuffer buffer;
-    VkDeviceMemory memory_1, memory_2;
-    std::string memory_name = "memory_name";
-
-    VkBufferCreateInfo buffer_create_info = vku::InitStructHelper();
-    buffer_create_info.usage = VK_BUFFER_USAGE_UNIFORM_BUFFER_BIT;
-    buffer_create_info.size = 1;
-
-    vk::CreateBuffer(device(), &buffer_create_info, nullptr, &buffer);
-
-    VkMemoryRequirements memRequirements;
-    vk::GetBufferMemoryRequirements(device(), buffer, &memRequirements);
-
-    VkMemoryAllocateInfo memory_allocate_info = vku::InitStructHelper();
-    memory_allocate_info.allocationSize = memRequirements.size;
-    memory_allocate_info.memoryTypeIndex = 0;
-
-    vk::AllocateMemory(device(), &memory_allocate_info, nullptr, &memory_1);
-    vk::AllocateMemory(device(), &memory_allocate_info, nullptr, &memory_2);
-
-    VkDebugMarkerObjectNameInfoEXT name_info = vku::InitStructHelper();
-    name_info.object = (uint64_t)memory_2;
-    name_info.objectType = VK_DEBUG_REPORT_OBJECT_TYPE_DEVICE_MEMORY_EXT;
-    name_info.pObjectName = memory_name.c_str();
-    vk::DebugMarkerSetObjectNameEXT(device(), &name_info);
-
-    vk::BindBufferMemory(device(), buffer, memory_1, 0);
-
-    // Test core_validation layer
-    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, memory_name);
-    vk::BindBufferMemory(device(), buffer, memory_2, 0);
-    m_errorMonitor->VerifyFound();
-
-    vk::FreeMemory(device(), memory_1, nullptr);
-    memory_1 = VK_NULL_HANDLE;
-    vk::FreeMemory(device(), memory_2, nullptr);
-    memory_2 = VK_NULL_HANDLE;
-    vk::DestroyBuffer(device(), buffer, nullptr);
-    buffer = VK_NULL_HANDLE;
-
-    VkCommandBuffer commandBuffer;
-    std::string commandBuffer_name = "command_buffer_name";
-    VkCommandPoolCreateInfo pool_create_info = vku::InitStructHelper();
-    pool_create_info.queueFamilyIndex = m_device->graphics_queue_node_index_;
-    pool_create_info.flags = VK_COMMAND_POOL_CREATE_RESET_COMMAND_BUFFER_BIT;
-    vkt::CommandPool command_pool_1(*m_device, pool_create_info);
-    vkt::CommandPool command_pool_2(*m_device, pool_create_info);
-
-    VkCommandBufferAllocateInfo command_buffer_allocate_info = vku::InitStructHelper();
-    command_buffer_allocate_info.commandPool = command_pool_1.handle();
-    command_buffer_allocate_info.commandBufferCount = 1;
-    command_buffer_allocate_info.level = VK_COMMAND_BUFFER_LEVEL_PRIMARY;
-    vk::AllocateCommandBuffers(device(), &command_buffer_allocate_info, &commandBuffer);
-
-    name_info.object = (uint64_t)commandBuffer;
-    name_info.objectType = VK_DEBUG_REPORT_OBJECT_TYPE_COMMAND_BUFFER_EXT;
-    name_info.pObjectName = commandBuffer_name.c_str();
-    vk::DebugMarkerSetObjectNameEXT(device(), &name_info);
-
-    VkCommandBufferBeginInfo cb_begin_Info = vku::InitStructHelper();
-    cb_begin_Info.flags = VK_COMMAND_BUFFER_USAGE_ONE_TIME_SUBMIT_BIT;
-    vk::BeginCommandBuffer(commandBuffer, &cb_begin_Info);
-
-    const VkRect2D scissor = {{-1, 0}, {16, 16}};
-    const VkRect2D scissors[] = {scissor, scissor};
-
-    // Test parameter_validation layer
-    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, commandBuffer_name);
-    vk::CmdSetScissor(commandBuffer, 0, 1, scissors);
-    m_errorMonitor->VerifyFound();
-
-    // Test object_tracker layer
-    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, commandBuffer_name);
-    vk::FreeCommandBuffers(device(), command_pool_2.handle(), 1, &commandBuffer);
-    m_errorMonitor->VerifyFound();
-}
-
-TEST_F(VkLayerTest, DebugUtilsNameTest) {
-    TEST_DESCRIPTION("Ensure debug utils object names are printed in debug messenger output");
-
-    AddRequiredExtensions(VK_EXT_DEBUG_UTILS_EXTENSION_NAME);
-
-    RETURN_IF_SKIP(InitFramework());
-    RETURN_IF_SKIP(InitState());
-
-    if (IsPlatformMockICD()) {
-        GTEST_SKIP() << "Skipping object naming test with MockICD.";
-    }
-
-    DebugUtilsLabelCheckData callback_data;
-    auto empty_callback = [](const VkDebugUtilsMessengerCallbackDataEXT *pCallbackData, DebugUtilsLabelCheckData *data) {
-        data->count++;
-    };
-    callback_data.count = 0;
-    callback_data.callback = empty_callback;
-
-    VkDebugUtilsMessengerCreateInfoEXT callback_create_info = vku::InitStructHelper();
-    callback_create_info.messageSeverity =
-        VK_DEBUG_UTILS_MESSAGE_SEVERITY_ERROR_BIT_EXT | VK_DEBUG_UTILS_MESSAGE_SEVERITY_WARNING_BIT_EXT;
-    callback_create_info.messageType = VK_DEBUG_UTILS_MESSAGE_TYPE_VALIDATION_BIT_EXT;
-    callback_create_info.pfnUserCallback = DebugUtilsCallback;
-    callback_create_info.pUserData = &callback_data;
-    VkDebugUtilsMessengerEXT my_messenger = VK_NULL_HANDLE;
-    vk::CreateDebugUtilsMessengerEXT(instance(), &callback_create_info, nullptr, &my_messenger);
-
-    VkBuffer buffer;
-    VkDeviceMemory memory_1, memory_2;
-    std::string memory_name = "memory_name";
-
-    VkBufferCreateInfo buffer_create_info = vku::InitStructHelper();
-    buffer_create_info.usage = VK_BUFFER_USAGE_UNIFORM_BUFFER_BIT;
-    buffer_create_info.size = 1;
-
-    vk::CreateBuffer(device(), &buffer_create_info, nullptr, &buffer);
-
-    VkMemoryRequirements memRequirements;
-    vk::GetBufferMemoryRequirements(device(), buffer, &memRequirements);
-
-    VkMemoryAllocateInfo memory_allocate_info = vku::InitStructHelper();
-    memory_allocate_info.allocationSize = memRequirements.size;
-    memory_allocate_info.memoryTypeIndex = 0;
-
-    vk::AllocateMemory(device(), &memory_allocate_info, nullptr, &memory_1);
-    vk::AllocateMemory(device(), &memory_allocate_info, nullptr, &memory_2);
-
-    VkDebugUtilsObjectNameInfoEXT name_info = vku::InitStructHelper();
-    name_info.objectType = VK_OBJECT_TYPE_DEVICE_MEMORY;
-    name_info.pObjectName = memory_name.c_str();
-
-    // Pass in bad handle make sure ObjectTracker catches it
-    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-VkDebugUtilsObjectNameInfoEXT-objectType-02590");
-    name_info.objectHandle = (uint64_t)0xcadecade;
-    vk::SetDebugUtilsObjectNameEXT(device(), &name_info);
-    m_errorMonitor->VerifyFound();
-
-    // Pass in null handle
-    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-vkSetDebugUtilsObjectNameEXT-pNameInfo-02588");
-    name_info.objectHandle = 0;
-    vk::SetDebugUtilsObjectNameEXT(device(), &name_info);
-    m_errorMonitor->VerifyFound();
-
-    // Pass in 'unknown' object type and see if parameter validation catches it
-    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-vkSetDebugUtilsObjectNameEXT-pNameInfo-02587");
-    name_info.objectHandle = (uint64_t)memory_2;
-    name_info.objectType = VK_OBJECT_TYPE_UNKNOWN;
-    vk::SetDebugUtilsObjectNameEXT(device(), &name_info);
-    m_errorMonitor->VerifyFound();
-
-    name_info.objectType = VK_OBJECT_TYPE_DEVICE_MEMORY;
-    vk::SetDebugUtilsObjectNameEXT(device(), &name_info);
-
-    vk::BindBufferMemory(device(), buffer, memory_1, 0);
-
-    // Test core_validation layer
-    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, memory_name);
-    vk::BindBufferMemory(device(), buffer, memory_2, 0);
-    m_errorMonitor->VerifyFound();
-
-    vk::FreeMemory(device(), memory_1, nullptr);
-    memory_1 = VK_NULL_HANDLE;
-    vk::FreeMemory(device(), memory_2, nullptr);
-    memory_2 = VK_NULL_HANDLE;
-    vk::DestroyBuffer(device(), buffer, nullptr);
-    buffer = VK_NULL_HANDLE;
-
-    VkCommandBuffer commandBuffer;
-    std::string commandBuffer_name = "command_buffer_name";
-    VkCommandPoolCreateInfo pool_create_info = vku::InitStructHelper();
-    pool_create_info.queueFamilyIndex = m_device->graphics_queue_node_index_;
-    pool_create_info.flags = VK_COMMAND_POOL_CREATE_RESET_COMMAND_BUFFER_BIT;
-    vkt::CommandPool command_pool_1(*m_device, pool_create_info);
-    vkt::CommandPool command_pool_2(*m_device, pool_create_info);
-
-    VkCommandBufferAllocateInfo command_buffer_allocate_info = vku::InitStructHelper();
-    command_buffer_allocate_info.commandPool = command_pool_1.handle();
-    command_buffer_allocate_info.commandBufferCount = 1;
-    command_buffer_allocate_info.level = VK_COMMAND_BUFFER_LEVEL_PRIMARY;
-    vk::AllocateCommandBuffers(device(), &command_buffer_allocate_info, &commandBuffer);
-
-    name_info.objectHandle = (uint64_t)commandBuffer;
-    name_info.objectType = VK_OBJECT_TYPE_COMMAND_BUFFER;
-    name_info.pObjectName = commandBuffer_name.c_str();
-    vk::SetDebugUtilsObjectNameEXT(device(), &name_info);
-
-    VkCommandBufferBeginInfo cb_begin_Info = vku::InitStructHelper();
-    cb_begin_Info.flags = VK_COMMAND_BUFFER_USAGE_ONE_TIME_SUBMIT_BIT;
-    vk::BeginCommandBuffer(commandBuffer, &cb_begin_Info);
-
-    const VkRect2D scissor = {{-1, 0}, {16, 16}};
-    const VkRect2D scissors[] = {scissor, scissor};
-
-    VkDebugUtilsLabelEXT command_label = vku::InitStructHelper();
-    command_label.pLabelName = "Command Label 0123";
-    command_label.color[0] = 0.;
-    command_label.color[1] = 1.;
-    command_label.color[2] = 2.;
-    command_label.color[3] = 3.0;
-    bool command_label_test = false;
-    auto command_label_callback = [command_label, &command_label_test](const VkDebugUtilsMessengerCallbackDataEXT *pCallbackData,
-                                                                       DebugUtilsLabelCheckData *data) {
-        data->count++;
-        command_label_test = false;
-        if (pCallbackData->cmdBufLabelCount == 1) {
-            command_label_test = pCallbackData->pCmdBufLabels[0] == command_label;
-        }
-    };
-    callback_data.callback = command_label_callback;
-
-    vk::CmdInsertDebugUtilsLabelEXT(commandBuffer, &command_label);
-    // Test parameter_validation layer
-    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, commandBuffer_name);
-    vk::CmdSetScissor(commandBuffer, 0, 1, scissors);
-    m_errorMonitor->VerifyFound();
-
-    // Check the label test
-    if (!command_label_test) {
-        ADD_FAILURE() << "Command label '" << command_label.pLabelName << "' not passed to callback.";
-    }
-
-    // Test object_tracker layer
-    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, commandBuffer_name);
-    vk::FreeCommandBuffers(device(), command_pool_2.handle(), 1, &commandBuffer);
-    m_errorMonitor->VerifyFound();
-
-    vk::DestroyDebugUtilsMessengerEXT(instance(), my_messenger, nullptr);
-}
-
-TEST_F(VkLayerTest, DebugUtilsParameterFlags) {
-    AddRequiredExtensions(VK_EXT_DEBUG_UTILS_EXTENSION_NAME);
-    RETURN_IF_SKIP(Init());
-
-    DebugUtilsLabelCheckData callback_data;
-    auto empty_callback = [](const VkDebugUtilsMessengerCallbackDataEXT *pCallbackData, DebugUtilsLabelCheckData *data) {
-        data->count++;
-    };
-    callback_data.count = 0;
-    callback_data.callback = empty_callback;
-
-    VkDebugUtilsMessengerCreateInfoEXT callback_create_info = vku::InitStructHelper();
-    callback_create_info.messageSeverity = 0;
-    callback_create_info.messageType = VK_DEBUG_UTILS_MESSAGE_TYPE_VALIDATION_BIT_EXT;
-    callback_create_info.pfnUserCallback = DebugUtilsCallback;
-    callback_create_info.pUserData = &callback_data;
-    VkDebugUtilsMessengerEXT my_messenger = VK_NULL_HANDLE;
-    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-VkDebugUtilsMessengerCreateInfoEXT-messageSeverity-requiredbitmask");
-    vk::CreateDebugUtilsMessengerEXT(instance(), &callback_create_info, nullptr, &my_messenger);
-    m_errorMonitor->VerifyFound();
-}
-
 TEST_F(VkLayerTest, InvalidStructSType) {
     TEST_DESCRIPTION("Specify an invalid VkStructureType for a Vulkan structure's sType field");
 
@@ -1185,98 +879,6 @@ TEST_F(VkLayerTest, UnrecognizedValueMaxEnum) {
     m_errorMonitor->VerifyFound();
 }
 
-TEST_F(VkLayerTest, LeakAnObject) {
-    TEST_DESCRIPTION("Create a fence and destroy its device without first destroying the fence.");
-
-    RETURN_IF_SKIP(InitFramework());
-    if (!IsPlatformMockICD()) {
-        // This test leaks a fence (on purpose) and should not be run on a real driver
-        GTEST_SKIP() << "This test only runs on the mock ICD";
-    }
-
-    // Workaround for overzealous layers checking even the guaranteed 0th queue family
-    const auto q_props = vkt::PhysicalDevice(gpu()).queue_properties_;
-    ASSERT_TRUE(q_props.size() > 0);
-    ASSERT_TRUE(q_props[0].queueCount > 0);
-
-    const float q_priority[] = {1.0f};
-    VkDeviceQueueCreateInfo queue_ci = vku::InitStructHelper();
-    queue_ci.queueFamilyIndex = 0;
-    queue_ci.queueCount = 1;
-    queue_ci.pQueuePriorities = q_priority;
-
-    VkDeviceCreateInfo device_ci = vku::InitStructHelper();
-    device_ci.queueCreateInfoCount = 1;
-    device_ci.pQueueCreateInfos = &queue_ci;
-
-    VkDevice leaky_device;
-    ASSERT_EQ(VK_SUCCESS, vk::CreateDevice(gpu(), &device_ci, nullptr, &leaky_device));
-
-    const VkFenceCreateInfo fence_ci = vku::InitStructHelper();
-    VkFence leaked_fence;
-    ASSERT_EQ(VK_SUCCESS, vk::CreateFence(leaky_device, &fence_ci, nullptr, &leaked_fence));
-
-    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-vkDestroyDevice-device-05137");
-    vk::DestroyDevice(leaky_device, nullptr);
-    m_errorMonitor->VerifyFound();
-
-    // There's no way we can destroy the fence at this point. Even though DestroyDevice failed, the loader has already removed
-    // references to the device
-    m_errorMonitor->SetUnexpectedError("VUID-vkDestroyDevice-device-05137");
-    m_errorMonitor->SetUnexpectedError("VUID-vkDestroyInstance-instance-00629");
-}
-
-TEST_F(VkLayerTest, LeakABuffer) {
-    TEST_DESCRIPTION("Create a fence and destroy its device without first destroying the buffer.");
-
-    RETURN_IF_SKIP(InitFramework());
-    if (!IsPlatformMockICD()) {
-        // This test leaks a buffer (on purpose) and should not be run on a real driver
-        GTEST_SKIP() << "This test only runs on the mock ICD";
-    }
-
-    // Workaround for overzealous layers checking even the guaranteed 0th queue family
-    const auto q_props = vkt::PhysicalDevice(gpu()).queue_properties_;
-    ASSERT_TRUE(q_props.size() > 0);
-    ASSERT_TRUE(q_props[0].queueCount > 0);
-
-    auto features = vkt::PhysicalDevice(gpu()).features();
-    if (!features.sparseBinding) {
-        GTEST_SKIP() << "Test requires unsupported sparseBinding feature";
-    }
-
-    const float q_priority[] = {1.0f};
-    VkDeviceQueueCreateInfo queue_ci = vku::InitStructHelper();
-    queue_ci.queueFamilyIndex = 0;
-    queue_ci.queueCount = 1;
-    queue_ci.pQueuePriorities = q_priority;
-
-    VkDeviceCreateInfo device_ci = vku::InitStructHelper();
-    device_ci.queueCreateInfoCount = 1;
-    device_ci.pQueueCreateInfos = &queue_ci;
-    device_ci.pEnabledFeatures = &features;
-
-    VkDevice leaky_device;
-    ASSERT_EQ(VK_SUCCESS, vk::CreateDevice(gpu(), &device_ci, nullptr, &leaky_device));
-
-    VkBufferCreateInfo buffer_create_info = vku::InitStructHelper();
-    buffer_create_info.usage = VK_BUFFER_USAGE_UNIFORM_BUFFER_BIT;
-    buffer_create_info.flags = VK_BUFFER_CREATE_SPARSE_BINDING_BIT;
-    buffer_create_info.size = 1;
-
-    VkBuffer buffer{};
-    ASSERT_EQ(VK_SUCCESS, vk::CreateBuffer(leaky_device, &buffer_create_info, nullptr, &buffer));
-
-    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-vkDestroyDevice-device-05137");
-    vk::DestroyDevice(leaky_device, nullptr);
-    m_errorMonitor->VerifyFound();
-
-    // There's no way we can destroy the buffer at this point.
-    // Even though DestroyDevice failed, the loader has already removed references to the device
-    m_errorMonitor->SetUnexpectedError("VUID-vkDestroyDevice-device-05137");
-    m_errorMonitor->SetUnexpectedError("VUID-vkDestroyInstance-instance-00629");
-}
-
 TEST_F(VkLayerTest, UseObjectWithWrongDevice) {
     TEST_DESCRIPTION(
         "Try to destroy a render pass object using a device other than the one it was created on. This should generate a distinct "
@@ -1368,45 +970,6 @@ TEST_F(VkLayerTest, InvalidAllocationCallbacks) {
         vk::CreateCommandPool(device(), &cpci, &allocator, &cmdPool);
         m_errorMonitor->VerifyFound();
     }
-}
-
-TEST_F(VkLayerTest, MismatchedQueueFamiliesOnSubmit) {
-    TEST_DESCRIPTION(
-        "Submit command buffer created using one queue family and attempt to submit them on a queue created in a different queue "
-        "family.");
-
-    RETURN_IF_SKIP(Init());  // assumes it initializes all queue families on vk::CreateDevice
-
-    // This test is meaningless unless we have multiple queue families
-    auto queue_family_properties = m_device->phy().queue_properties_;
-    std::vector<uint32_t> queue_families;
-    for (uint32_t i = 0; i < queue_family_properties.size(); ++i)
-        if (queue_family_properties[i].queueCount > 0) queue_families.push_back(i);
-
-    if (queue_families.size() < 2) {
-        GTEST_SKIP() << "Device only has one queue family";
-    }
-
-    const uint32_t queue_family = queue_families[0];
-
-    const uint32_t other_queue_family = queue_families[1];
-    VkQueue other_queue;
-    vk::GetDeviceQueue(m_device->device(), other_queue_family, 0, &other_queue);
-
-    vkt::CommandPool cmd_pool(*m_device, queue_family);
-    vkt::CommandBuffer cmd_buff(m_device, &cmd_pool);
-
-    cmd_buff.begin();
-    cmd_buff.end();
-
-    // Submit on the wrong queue
-    VkSubmitInfo submit_info = vku::InitStructHelper();
-    submit_info.commandBufferCount = 1;
-    submit_info.pCommandBuffers = &cmd_buff.handle();
-
-    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-vkQueueSubmit-pCommandBuffers-00074");
-    vk::QueueSubmit(other_queue, 1, &submit_info, VK_NULL_HANDLE);
-    m_errorMonitor->VerifyFound();
 }
 
 TEST_F(VkLayerTest, DeviceFeature2AndVertexAttributeDivisorExtensionUnenabled) {
@@ -1658,49 +1221,6 @@ TEST_F(VkLayerTest, UnclosedAndDuplicateQueries) {
     m_errorMonitor->VerifyFound();
 }
 
-TEST_F(VkLayerTest, StageMaskHost) {
-    TEST_DESCRIPTION("Test invalid usage of VK_PIPELINE_STAGE_HOST_BIT.");
-    RETURN_IF_SKIP(Init());
-    InitRenderTarget();
-
-    vkt::Event event(*m_device);
-    m_commandBuffer->begin();
-
-    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-vkCmdSetEvent-stageMask-01149");
-    vk::CmdSetEvent(m_commandBuffer->handle(), event.handle(), VK_PIPELINE_STAGE_HOST_BIT);
-    m_errorMonitor->VerifyFound();
-
-    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-vkCmdResetEvent-stageMask-01153");
-    vk::CmdResetEvent(m_commandBuffer->handle(), event.handle(), VK_PIPELINE_STAGE_HOST_BIT);
-    m_errorMonitor->VerifyFound();
-
-    m_commandBuffer->end();
-
-    vkt::Semaphore semaphore(*m_device);
-    ASSERT_TRUE(semaphore.initialized());
-
-    VkPipelineStageFlags stage_flags = VK_PIPELINE_STAGE_HOST_BIT;
-    VkSubmitInfo submit_info = vku::InitStructHelper();
-
-    // Signal the semaphore so the next test can wait on it.
-    submit_info.signalSemaphoreCount = 1;
-    submit_info.pSignalSemaphores = &semaphore.handle();
-    vk::QueueSubmit(m_default_queue->handle(), 1, &submit_info, VK_NULL_HANDLE);
-
-    submit_info.signalSemaphoreCount = 0;
-    submit_info.pSignalSemaphores = nullptr;
-    submit_info.waitSemaphoreCount = 1;
-    submit_info.pWaitSemaphores = &semaphore.handle();
-    submit_info.pWaitDstStageMask = &stage_flags;
-
-    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-VkSubmitInfo-pWaitDstStageMask-00078");
-    vk::QueueSubmit(m_default_queue->handle(), 1, &submit_info, VK_NULL_HANDLE);
-    m_errorMonitor->VerifyFound();
-
-    // Need to ensure semaphore is not in use before the test ends and it gets destroyed
-    m_default_queue->wait();
-}
-
 TEST_F(VkLayerTest, ExecuteUnrecordedCB) {
     TEST_DESCRIPTION("Attempt vkQueueSubmit with a CB in the initial state");
 
@@ -1780,90 +1300,6 @@ TEST_F(VkLayerTest, ApiVersion1_1AndNegativeViewport) {
     // not use the LunarG loader (e.g. Android) will not see the message and the test will fail.
     m_errorMonitor->SetUnexpectedError("Failed to create device chain.");
     vk::CreateDevice(gpu(), &device_create_info, NULL, &testDevice);
-    m_errorMonitor->VerifyFound();
-}
-
-TEST_F(VkLayerTest, ResetEventThenSet) {
-    TEST_DESCRIPTION("Reset an event then set it after the reset has been submitted.");
-
-    RETURN_IF_SKIP(Init());
-    VkEventCreateInfo event_create_info = vku::InitStructHelper();
-    vkt::Event event(*m_device, event_create_info);
-
-    VkCommandPoolCreateInfo pool_create_info = vku::InitStructHelper();
-    pool_create_info.queueFamilyIndex = m_device->graphics_queue_node_index_;
-    pool_create_info.flags = VK_COMMAND_POOL_CREATE_RESET_COMMAND_BUFFER_BIT;
-    vkt::CommandPool command_pool(*m_device, pool_create_info);
-
-    VkCommandBuffer command_buffer;
-    VkCommandBufferAllocateInfo command_buffer_allocate_info = vku::InitStructHelper();
-    command_buffer_allocate_info.commandPool = command_pool.handle();
-    command_buffer_allocate_info.commandBufferCount = 1;
-    command_buffer_allocate_info.level = VK_COMMAND_BUFFER_LEVEL_PRIMARY;
-    vk::AllocateCommandBuffers(m_device->device(), &command_buffer_allocate_info, &command_buffer);
-
-    VkQueue queue = VK_NULL_HANDLE;
-    vk::GetDeviceQueue(m_device->device(), m_device->graphics_queue_node_index_, 0, &queue);
-
-    {
-        VkCommandBufferBeginInfo begin_info = vku::InitStructHelper();
-        vk::BeginCommandBuffer(command_buffer, &begin_info);
-
-        vk::CmdResetEvent(command_buffer, event.handle(), VK_PIPELINE_STAGE_ALL_COMMANDS_BIT);
-        vk::EndCommandBuffer(command_buffer);
-    }
-    {
-        VkSubmitInfo submit_info = vku::InitStructHelper();
-        submit_info.commandBufferCount = 1;
-        submit_info.pCommandBuffers = &command_buffer;
-        submit_info.signalSemaphoreCount = 0;
-        submit_info.pSignalSemaphores = nullptr;
-        vk::QueueSubmit(queue, 1, &submit_info, VK_NULL_HANDLE);
-    }
-    {
-        m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "UNASSIGNED-vkSetEvent-QueueForwardProgress");
-        vk::SetEvent(m_device->device(), event.handle());
-        m_errorMonitor->VerifyFound();
-    }
-
-    vk::QueueWaitIdle(queue);
-
-    vk::FreeCommandBuffers(m_device->device(), command_pool.handle(), 1, &command_buffer);
-}
-
-TEST_F(VkLayerTest, FreeCommandBuffersNull) {
-    TEST_DESCRIPTION("Can pass NULL for vkFreeCommandBuffers");
-    RETURN_IF_SKIP(Init());
-
-    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-vkFreeCommandBuffers-pCommandBuffers-00048");
-    vk::FreeCommandBuffers(m_device->device(), m_commandPool->handle(), 2, nullptr);
-    m_errorMonitor->VerifyFound();
-
-    VkCommandBuffer invalid_cb = CastToHandle<VkCommandBuffer, uintptr_t>(0xbaadbeef);
-    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-vkFreeCommandBuffers-pCommandBuffers-00048");
-    vk::FreeCommandBuffers(m_device->device(), m_commandPool->handle(), 1, &invalid_cb);
-    m_errorMonitor->VerifyFound();
-}
-
-TEST_F(VkLayerTest, FreeDescriptorSetsNull) {
-    TEST_DESCRIPTION("Can pass NULL for vkFreeDescriptorSets");
-    RETURN_IF_SKIP(Init());
-
-    VkDescriptorPoolSize ds_type_count = {VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER_DYNAMIC, 1};
-    VkDescriptorPoolCreateInfo ds_pool_ci = vku::InitStructHelper();
-    ds_pool_ci.maxSets = 1;
-    ds_pool_ci.poolSizeCount = 1;
-    ds_pool_ci.flags = VK_DESCRIPTOR_POOL_CREATE_FREE_DESCRIPTOR_SET_BIT;
-    ds_pool_ci.pPoolSizes = &ds_type_count;
-    vkt::DescriptorPool ds_pool(*m_device, ds_pool_ci);
-
-    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-vkFreeDescriptorSets-pDescriptorSets-00310");
-    vk::FreeDescriptorSets(m_device->device(), ds_pool.handle(), 2, nullptr);
-    m_errorMonitor->VerifyFound();
-
-    VkDescriptorSet invalid_set = CastToHandle<VkDescriptorSet, uintptr_t>(0xbaadbeef);
-    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-vkFreeDescriptorSets-pDescriptorSets-00310");
-    vk::FreeDescriptorSets(m_device->device(), ds_pool.handle(), 1, &invalid_set);
     m_errorMonitor->VerifyFound();
 }
 
@@ -2092,100 +1528,6 @@ TEST_F(VkLayerTest, ValidateArrayLength) {
     command_buffer.end();
 }
 
-TEST_F(VkLayerTest, InvalidSpirvExtension) {
-    TEST_DESCRIPTION("Use an invalid SPIR-V extension in OpExtension.");
-
-    SetTargetApiVersion(VK_API_VERSION_1_2);
-    RETURN_IF_SKIP(Init());
-
-    InitRenderTarget();
-
-    const char *vertex_source = R"spirv(
-               OpCapability Shader
-          %1 = OpExtInstImport "GLSL.std.450"
-               OpMemoryModel Logical GLSL450
-               OpEntryPoint Vertex %4 "main"
-               OpSource GLSL 450
-               OpExtension "GL_EXT_scalar_block_layout"
-               OpName %4 "main"
-          %2 = OpTypeVoid
-          %3 = OpTypeFunction %2
-          %4 = OpFunction %2 None %3
-          %5 = OpLabel
-               OpReturn
-               OpFunctionEnd
-        )spirv";
-    VkShaderObj vs(this, vertex_source, VK_SHADER_STAGE_VERTEX_BIT, SPV_ENV_VULKAN_1_2, SPV_SOURCE_ASM_TRY);
-    m_errorMonitor->SetUnexpectedError("VUID-VkShaderModuleCreateInfo-pCode-08737");
-    if (vs.InitFromASMTry() != VK_SUCCESS) {
-        GTEST_SKIP() << "Failed to compile shader";
-    }
-    const VkShaderObj fs(this, kFragmentMinimalGlsl, VK_SHADER_STAGE_FRAGMENT_BIT);
-
-    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-VkShaderModuleCreateInfo-pCode-08741");
-    CreatePipelineHelper pipe(*this);
-    pipe.shader_stages_ = {vs.GetStageCreateInfo(), fs.GetStageCreateInfo()};
-    pipe.InitState();
-    pipe.CreateGraphicsPipeline();
-    m_errorMonitor->VerifyFound();
-}
-
-TEST_F(VkLayerTest, ValidateViewportStateScissorOverflow) {
-    TEST_DESCRIPTION("Validate sum of offset and width of viewport state scissor");
-
-    RETURN_IF_SKIP(Init());
-    InitRenderTarget();
-
-    VkViewport viewport = {0.0f, 0.0f, 64.0f, 64.0f, 0.0f, 1.0f};
-    VkRect2D scissor_x = {{vvl::kI32Max / 2, 0}, {vvl::kI32Max / 2 + 64, 64}};
-    VkRect2D scissor_y = {{0, vvl::kI32Max / 2}, {64, vvl::kI32Max / 2 + 64}};
-
-    const auto break_vp_x = [&](CreatePipelineHelper &helper) {
-        helper.vp_state_ci_.viewportCount = 1;
-        helper.vp_state_ci_.pViewports = &viewport;
-        helper.vp_state_ci_.scissorCount = 1;
-        helper.vp_state_ci_.pScissors = &scissor_x;
-    };
-    CreatePipelineHelper::OneshotTest(*this, break_vp_x, kErrorBit,
-                                      vector<std::string>({"VUID-VkPipelineViewportStateCreateInfo-offset-02822"}));
-
-    const auto break_vp_y = [&](CreatePipelineHelper &helper) {
-        helper.vp_state_ci_.viewportCount = 1;
-        helper.vp_state_ci_.pViewports = &viewport;
-        helper.vp_state_ci_.scissorCount = 1;
-        helper.vp_state_ci_.pScissors = &scissor_y;
-    };
-    CreatePipelineHelper::OneshotTest(*this, break_vp_y, kErrorBit,
-                                      vector<std::string>({"VUID-VkPipelineViewportStateCreateInfo-offset-02823"}));
-}
-
-TEST_F(VkLayerTest, ValidateViewportStateScissorNegative) {
-    TEST_DESCRIPTION("Validate offset of viewport state scissor");
-
-    RETURN_IF_SKIP(Init());
-    InitRenderTarget();
-
-    VkViewport viewport = {0.0f, 0.0f, 64.0f, 64.0f, 0.0f, 1.0f};
-    VkRect2D scissor_x = {{-64, 0}, {256, 256}};
-    VkRect2D scissor_y = {{0, -64}, {256, 256}};
-
-    const auto break_vp_x = [&](CreatePipelineHelper &helper) {
-        helper.vp_state_ci_.viewportCount = 1;
-        helper.vp_state_ci_.pViewports = &viewport;
-        helper.vp_state_ci_.scissorCount = 1;
-        helper.vp_state_ci_.pScissors = &scissor_x;
-    };
-    CreatePipelineHelper::OneshotTest(*this, break_vp_x, kErrorBit, "VUID-VkPipelineViewportStateCreateInfo-x-02821");
-
-    const auto break_vp_y = [&](CreatePipelineHelper &helper) {
-        helper.vp_state_ci_.viewportCount = 1;
-        helper.vp_state_ci_.pViewports = &viewport;
-        helper.vp_state_ci_.scissorCount = 1;
-        helper.vp_state_ci_.pScissors = &scissor_y;
-    };
-    CreatePipelineHelper::OneshotTest(*this, break_vp_y, kErrorBit, "VUID-VkPipelineViewportStateCreateInfo-x-02821");
-}
-
 TEST_F(VkLayerTest, DuplicatePhysicalDevices) {
     TEST_DESCRIPTION("Duplicated physical devices in DeviceGroupDeviceCreateInfo.");
     SetTargetApiVersion(VK_API_VERSION_1_1);
@@ -2356,55 +1698,6 @@ TEST_F(VkLayerTest, Features12AndppEnabledExtensionNames) {
     m_errorMonitor->VerifyFound();
 }
 
-TEST_F(VkLayerTest, ImageSubresourceOverlapBetweenCurrentRenderPassAndDescriptorSets) {
-    TEST_DESCRIPTION("Validate if attachments in render pass and descriptor set use the same image subresources");
-
-    RETURN_IF_SKIP(Init());
-    InitRenderTarget();
-
-    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "UNASSIGNED-CoreValidation-DrawState-InvalidRenderpass");
-    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-VkRenderPassBeginInfo-renderPass-00904");
-
-    const uint32_t width = 16;
-    const uint32_t height = 16;
-    const VkFormat format = VK_FORMAT_R8G8B8A8_UNORM;
-
-    VkAttachmentReference attach_ref = {};
-    attach_ref.attachment = 0;
-    attach_ref.layout = VK_IMAGE_LAYOUT_GENERAL;
-    VkSubpassDescription subpass = {};
-    subpass.pipelineBindPoint = VK_PIPELINE_BIND_POINT_GRAPHICS;
-    subpass.colorAttachmentCount = 1;
-    subpass.pColorAttachments = &attach_ref;
-
-    VkAttachmentDescription attach_desc = {};
-    attach_desc.format = format;
-    attach_desc.samples = VK_SAMPLE_COUNT_1_BIT;
-    attach_desc.loadOp = VK_ATTACHMENT_LOAD_OP_CLEAR;
-    attach_desc.storeOp = VK_ATTACHMENT_STORE_OP_STORE;
-    attach_desc.stencilLoadOp = VK_ATTACHMENT_LOAD_OP_DONT_CARE;
-    attach_desc.stencilStoreOp = VK_ATTACHMENT_STORE_OP_DONT_CARE;
-    attach_desc.initialLayout = VK_IMAGE_LAYOUT_UNDEFINED;
-    attach_desc.finalLayout = VK_IMAGE_LAYOUT_GENERAL;
-    VkAttachmentDescription attach_desc2[] = {attach_desc, attach_desc};
-
-    VkRenderPassCreateInfo rpci = vku::InitStructHelper();
-    rpci.subpassCount = 1;
-    rpci.pSubpasses = &subpass;
-    rpci.attachmentCount = 2;
-    rpci.pAttachments = attach_desc2;
-
-    vkt::RenderPass render_pass(*m_device, rpci);
-
-    VkClearValue clear_values[2] = {m_renderPassClearValues[0], m_renderPassClearValues[0]};
-
-    m_commandBuffer->begin();
-    m_commandBuffer->BeginRenderPass(render_pass.handle(), framebuffer(), width, height, 2, clear_values);
-    m_commandBuffer->end();
-
-    m_errorMonitor->VerifyFound();
-}
-
 TEST_F(VkLayerTest, ZeroBitmask) {
     TEST_DESCRIPTION("Test a reserved flags field set to a non-zero value");
 
@@ -2417,332 +1710,6 @@ TEST_F(VkLayerTest, ZeroBitmask) {
     vk::CreateSemaphore(m_device->device(), &semaphore_ci, nullptr, &semaphore);
     m_errorMonitor->VerifyFound();
 }
-
-TEST_F(VkLayerTest, InstanceCreateEnumeratePortability) {
-    TEST_DESCRIPTION("Validate creating instances with VK_INSTANCE_CREATE_ENUMERATE_PORTABILITY_BIT_KHR.");
-
-    auto ici = GetInstanceCreateInfo();
-    ici.flags = VK_INSTANCE_CREATE_ENUMERATE_PORTABILITY_BIT_KHR;
-    ici.enabledExtensionCount = 1;
-
-    VkInstance local_instance;
-
-    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-VkInstanceCreateInfo-flags-06559");
-    vk::CreateInstance(&ici, nullptr, &local_instance);
-    m_errorMonitor->VerifyFound();
-
-    if (InstanceExtensionSupported(VK_KHR_PORTABILITY_ENUMERATION_EXTENSION_NAME)) {
-        std::vector<const char *> enabled_extensions = {VK_KHR_PORTABILITY_ENUMERATION_EXTENSION_NAME,
-                                                        VK_EXT_DEBUG_UTILS_EXTENSION_NAME};
-        ici.enabledExtensionCount = static_cast<uint32_t>(enabled_extensions.size());
-        ici.ppEnabledExtensionNames = enabled_extensions.data();
-
-        ASSERT_EQ(VK_SUCCESS, vk::CreateInstance(&ici, nullptr, &local_instance));
-        vk::DestroyInstance(local_instance, nullptr);
-    }
-}
-
-#ifdef VK_USE_PLATFORM_METAL_EXT
-TEST_F(VkLayerTest, ExportMetalObjects) {
-    TEST_DESCRIPTION("Test VK_EXT_metal_objects VUIDs");
-
-    AddRequiredExtensions(VK_KHR_GET_PHYSICAL_DEVICE_PROPERTIES_2_EXTENSION_NAME);
-    AddRequiredExtensions(VK_EXT_METAL_OBJECTS_EXTENSION_NAME);
-    AddOptionalExtensions(VK_KHR_SAMPLER_YCBCR_CONVERSION_EXTENSION_NAME);
-
-    RETURN_IF_SKIP(InitFramework());
-    const bool ycbcr_conversion_extension = IsExtensionsEnabled(VK_KHR_SAMPLER_YCBCR_CONVERSION_EXTENSION_NAME);
-    VkPhysicalDevicePortabilitySubsetFeaturesKHR portability_features = vku::InitStructHelper();
-    auto features2 = GetPhysicalDeviceFeatures2(portability_features);
-
-    if (ycbcr_conversion_extension) {
-        VkPhysicalDeviceSamplerYcbcrConversionFeatures ycbcr_features = vku::InitStructHelper();
-        ycbcr_features.samplerYcbcrConversion = VK_TRUE;
-        portability_features.pNext = &ycbcr_features;
-    }
-
-    RETURN_IF_SKIP(InitState(nullptr, &features2));
-
-    VkExportMetalObjectCreateInfoEXT metal_object_create_info = vku::InitStructHelper();
-    auto instance_ci = GetInstanceCreateInfo();
-    metal_object_create_info.exportObjectType = VK_EXPORT_METAL_OBJECT_TYPE_METAL_SHARED_EVENT_BIT_EXT;
-    metal_object_create_info.pNext = instance_ci.pNext;
-    instance_ci.pNext = &metal_object_create_info;
-
-    VkInstance instance = {};
-    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-VkInstanceCreateInfo-pNext-06779");
-    vk::CreateInstance(&instance_ci, nullptr, &instance);
-    m_errorMonitor->VerifyFound();
-    metal_object_create_info.pNext = nullptr;
-
-    VkMemoryAllocateInfo alloc_info = vku::InitStructHelper();
-    alloc_info.pNext = &metal_object_create_info;
-    alloc_info.allocationSize = 1024;
-    VkDeviceMemory memory;
-    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-VkMemoryAllocateInfo-pNext-06780");
-    vk::AllocateMemory(device(), &alloc_info, nullptr, &memory);
-    m_errorMonitor->VerifyFound();
-
-    VkImageCreateInfo ici = vku::InitStructHelper();
-    ici.imageType = VK_IMAGE_TYPE_2D;
-    ici.format = VK_FORMAT_B8G8R8A8_UNORM;
-    ici.extent = {128, 128, 1};
-    ici.mipLevels = 1;
-    ici.arrayLayers = 1;
-    ici.samples = VK_SAMPLE_COUNT_1_BIT;
-    ici.tiling = VK_IMAGE_TILING_LINEAR;
-    ici.usage = VK_IMAGE_USAGE_TRANSFER_SRC_BIT;
-    ici.sharingMode = VK_SHARING_MODE_EXCLUSIVE;
-    ici.initialLayout = VK_IMAGE_LAYOUT_UNDEFINED;
-    ici.pNext = &metal_object_create_info;
-    CreateImageTest(*this, &ici, "VUID-VkImageCreateInfo-pNext-06783");
-
-    VkImportMetalTextureInfoEXT import_metal_texture_info = vku::InitStructHelper();
-    import_metal_texture_info.plane = VK_IMAGE_ASPECT_COLOR_BIT;
-    ici.pNext = &import_metal_texture_info;
-    ici.format = VK_FORMAT_G8_B8R8_2PLANE_420_UNORM;
-    CreateImageTest(*this, &ici, "VUID-VkImageCreateInfo-pNext-06784");
-
-    ici.format = VK_FORMAT_B8G8R8A8_UNORM;
-    import_metal_texture_info.plane = VK_IMAGE_ASPECT_PLANE_1_BIT;
-    CreateImageTest(*this, &ici, "VUID-VkImageCreateInfo-pNext-06785");
-
-    ici.format = VK_FORMAT_G8_B8R8_2PLANE_420_UNORM;
-    import_metal_texture_info.plane = VK_IMAGE_ASPECT_PLANE_2_BIT;
-    CreateImageTest(*this, &ici, "VUID-VkImageCreateInfo-pNext-06786");
-
-    uint32_t queue_family_index = 0;
-    VkBufferCreateInfo buffer_create_info = vku::InitStructHelper();
-    buffer_create_info.size = 1024;
-    buffer_create_info.usage = VK_BUFFER_USAGE_STORAGE_TEXEL_BUFFER_BIT;
-    buffer_create_info.queueFamilyIndexCount = 1;
-    buffer_create_info.pQueueFamilyIndices = &queue_family_index;
-
-    vkt::Buffer buffer(*m_device, buffer_create_info);
-    VkBufferViewCreateInfo buff_view_ci = vku::InitStructHelper();
-    buff_view_ci.buffer = buffer.handle();
-    buff_view_ci.format = VK_FORMAT_B8G8R8A8_UNORM;
-    buff_view_ci.range = VK_WHOLE_SIZE;
-    buff_view_ci.pNext = &metal_object_create_info;
-    CreateBufferViewTest(*this, &buff_view_ci, {"VUID-VkBufferViewCreateInfo-pNext-06782"});
-
-    VkImageObj image_obj(m_device);
-    image_obj.Init(256, 256, 1, VK_FORMAT_B8G8R8A8_UNORM, VK_IMAGE_USAGE_STORAGE_BIT);
-    VkImageViewCreateInfo ivci = vku::InitStructHelper();
-    ivci.image = image_obj.handle();
-    ivci.viewType = VK_IMAGE_VIEW_TYPE_2D;
-    ivci.format = VK_FORMAT_B8G8R8A8_UNORM;
-    ivci.subresourceRange.layerCount = 1;
-    ivci.subresourceRange.baseMipLevel = 0;
-    ivci.subresourceRange.levelCount = 1;
-    ivci.subresourceRange.aspectMask = VK_IMAGE_ASPECT_COLOR_BIT;
-    ivci.pNext = &metal_object_create_info;
-    CreateImageViewTest(*this, &ivci, "VUID-VkImageViewCreateInfo-pNext-06787");
-
-    VkSemaphoreCreateInfo sem_info = vku::InitStructHelper();
-    sem_info.pNext = &metal_object_create_info;
-    VkSemaphore semaphore;
-    metal_object_create_info.exportObjectType = VK_EXPORT_METAL_OBJECT_TYPE_METAL_BUFFER_BIT_EXT;
-    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-VkSemaphoreCreateInfo-pNext-06789");
-    vk::CreateSemaphore(device(), &sem_info, NULL, &semaphore);
-    m_errorMonitor->VerifyFound();
-
-    VkEventCreateInfo event_info = vku::InitStructHelper();
-    if (portability_features.events) {
-        event_info.pNext = &metal_object_create_info;
-        VkEvent event;
-        m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-VkEventCreateInfo-pNext-06790");
-        vk::CreateEvent(device(), &event_info, nullptr, &event);
-        m_errorMonitor->VerifyFound();
-    }
-
-    VkExportMetalObjectsInfoEXT export_metal_objects_info = vku::InitStructHelper();
-    VkExportMetalDeviceInfoEXT metal_device_info = vku::InitStructHelper();
-    VkExportMetalCommandQueueInfoEXT metal_command_queue_info = vku::InitStructHelper();
-    metal_command_queue_info.queue = m_default_queue->handle();
-    export_metal_objects_info.pNext = &metal_device_info;
-    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-VkExportMetalObjectsInfoEXT-pNext-06791");
-    vk::ExportMetalObjectsEXT(m_device->handle(), &export_metal_objects_info);
-    m_errorMonitor->VerifyFound();
-
-    export_metal_objects_info.pNext = &metal_command_queue_info;
-    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-VkExportMetalObjectsInfoEXT-pNext-06792");
-    vk::ExportMetalObjectsEXT(m_device->handle(), &export_metal_objects_info);
-    m_errorMonitor->VerifyFound();
-
-    alloc_info.pNext = nullptr;
-    VkResult err = vk::AllocateMemory(device(), &alloc_info, nullptr, &memory);
-    ASSERT_EQ(VK_SUCCESS, err);
-    VkExportMetalBufferInfoEXT metal_buffer_info = vku::InitStructHelper();
-    metal_buffer_info.memory = memory;
-    export_metal_objects_info.pNext = &metal_buffer_info;
-    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-VkExportMetalObjectsInfoEXT-pNext-06793");
-    vk::ExportMetalObjectsEXT(m_device->handle(), &export_metal_objects_info);
-    m_errorMonitor->VerifyFound();
-    vk::FreeMemory(device(), memory, nullptr);
-
-    VkExportMetalObjectCreateInfoEXT export_metal_object_create_info = vku::InitStructHelper();
-    export_metal_object_create_info.exportObjectType = VK_EXPORT_METAL_OBJECT_TYPE_METAL_TEXTURE_BIT_EXT;
-    ici.pNext = &export_metal_object_create_info;
-    VkImageObj export_image_obj(m_device);
-    export_image_obj.Init(ici);
-    vkt::BufferView export_buffer_view;
-    buff_view_ci.pNext = &export_metal_object_create_info;
-    export_buffer_view.init(*m_device, buff_view_ci);
-    VkExportMetalTextureInfoEXT metal_texture_info = vku::InitStructHelper();
-    metal_texture_info.bufferView = export_buffer_view.handle();
-    metal_texture_info.image = export_image_obj.handle();
-    metal_texture_info.plane = VK_IMAGE_ASPECT_PLANE_0_BIT;
-    export_metal_objects_info.pNext = &metal_texture_info;
-
-    // Only one of image, bufferView, imageView
-    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-VkExportMetalObjectsInfoEXT-pNext-06794");
-    vk::ExportMetalObjectsEXT(m_device->handle(), &export_metal_objects_info);
-    m_errorMonitor->VerifyFound();
-
-    // Image not created with struct in pNext
-    metal_texture_info.bufferView = VK_NULL_HANDLE;
-    metal_texture_info.image = image_obj.handle();
-    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-VkExportMetalObjectsInfoEXT-pNext-06795");
-    vk::ExportMetalObjectsEXT(m_device->handle(), &export_metal_objects_info);
-    m_errorMonitor->VerifyFound();
-
-    metal_texture_info.image = VK_NULL_HANDLE;
-    auto image_view_ci = image_obj.BasicViewCreatInfo();
-    vkt::ImageView image_view_no_struct(*m_device, image_view_ci);
-    metal_texture_info.imageView = image_view_no_struct.handle();
-    // ImageView not created with struct in pNext
-    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-VkExportMetalObjectsInfoEXT-pNext-06796");
-    vk::ExportMetalObjectsEXT(m_device->handle(), &export_metal_objects_info);
-    m_errorMonitor->VerifyFound();
-
-    buff_view_ci.pNext = nullptr;
-    vkt::BufferView buffer_view_no_struct;
-    buffer_view_no_struct.init(*m_device, buff_view_ci);
-    metal_texture_info.imageView = VK_NULL_HANDLE;
-    metal_texture_info.bufferView = buffer_view_no_struct.handle();
-    // BufferView not created with struct in pNext
-    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-VkExportMetalObjectsInfoEXT-pNext-06797");
-    vk::ExportMetalObjectsEXT(m_device->handle(), &export_metal_objects_info);
-    m_errorMonitor->VerifyFound();
-
-    metal_texture_info.bufferView = VK_NULL_HANDLE;
-    metal_texture_info.image = export_image_obj.handle();
-    metal_texture_info.plane = VK_IMAGE_ASPECT_COLOR_BIT;
-    // metal_texture_info.plane not plane 0, 1 or 2
-    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-VkExportMetalObjectsInfoEXT-pNext-06798");
-    vk::ExportMetalObjectsEXT(m_device->handle(), &export_metal_objects_info);
-    m_errorMonitor->VerifyFound();
-
-    ici.format = VK_FORMAT_B8G8R8A8_UNORM;
-    VkImageObj single_plane_export_image_obj(m_device);
-    single_plane_export_image_obj.Init(ici);
-    metal_texture_info.plane = VK_IMAGE_ASPECT_PLANE_1_BIT;
-    metal_texture_info.image = single_plane_export_image_obj.handle();
-    // metal_texture_info.plane not plane_0 for single plane image
-    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-VkExportMetalObjectsInfoEXT-pNext-06799");
-    vk::ExportMetalObjectsEXT(m_device->handle(), &export_metal_objects_info);
-    m_errorMonitor->VerifyFound();
-
-    image_view_ci.pNext = &export_metal_object_create_info;
-    export_metal_object_create_info.exportObjectType = VK_EXPORT_METAL_OBJECT_TYPE_METAL_TEXTURE_BIT_EXT;
-    vkt::ImageView single_plane_export_image_view(*m_device, image_view_ci);
-    metal_texture_info.image = VK_NULL_HANDLE;
-    metal_texture_info.imageView = single_plane_export_image_view.handle();
-    // metal_texture_info.plane not plane_0 for single plane imageView
-    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-VkExportMetalObjectsInfoEXT-pNext-06801");
-    vk::ExportMetalObjectsEXT(m_device->handle(), &export_metal_objects_info);
-    m_errorMonitor->VerifyFound();
-
-    VkExportMetalIOSurfaceInfoEXT metal_iosurface_info = vku::InitStructHelper();
-    metal_iosurface_info.image = image_obj.handle();
-    export_metal_objects_info.pNext = &metal_iosurface_info;
-    // metal_iosurface_info.image not created with struct in pNext
-    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-VkExportMetalObjectsInfoEXT-pNext-06803");
-    vk::ExportMetalObjectsEXT(m_device->handle(), &export_metal_objects_info);
-    m_errorMonitor->VerifyFound();
-
-    VkExportMetalSharedEventInfoEXT metal_shared_event_info = vku::InitStructHelper();
-    export_metal_objects_info.pNext = &metal_shared_event_info;
-    // metal_shared_event_info event and semaphore both VK_NULL_HANDLE
-    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-VkExportMetalObjectsInfoEXT-pNext-06804");
-    vk::ExportMetalObjectsEXT(m_device->handle(), &export_metal_objects_info);
-    m_errorMonitor->VerifyFound();
-
-    sem_info.pNext = nullptr;
-    vkt::Semaphore semaphore_no_struct;
-    semaphore_no_struct.init(*m_device, sem_info);
-    metal_shared_event_info.semaphore = semaphore_no_struct.handle();
-    export_metal_objects_info.pNext = &metal_shared_event_info;
-    // Semaphore not created with struct in pNext
-    m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-VkExportMetalObjectsInfoEXT-pNext-06805");
-    vk::ExportMetalObjectsEXT(m_device->handle(), &export_metal_objects_info);
-    m_errorMonitor->VerifyFound();
-
-    if (portability_features.events) {
-        event_info.pNext = nullptr;
-        vkt::Event event_no_struct(*m_device, event_info);
-        metal_shared_event_info.event = event_no_struct.handle();
-        metal_shared_event_info.semaphore = VK_NULL_HANDLE;
-        // Event not created with struct in pNext
-        m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-VkExportMetalObjectsInfoEXT-pNext-06806");
-        vk::ExportMetalObjectsEXT(m_device->handle(), &export_metal_objects_info);
-        m_errorMonitor->VerifyFound();
-    }
-
-    const VkFormat mp_format = VK_FORMAT_G8_B8R8_2PLANE_420_UNORM;
-    if (FormatIsSupported(gpu(), mp_format)) {
-        export_metal_object_create_info = vku::InitStructHelper();
-        export_metal_object_create_info.exportObjectType = VK_EXPORT_METAL_OBJECT_TYPE_METAL_TEXTURE_BIT_EXT;
-        ici.format = mp_format;
-        ici.tiling = VK_IMAGE_TILING_OPTIMAL;
-        ici.usage = VK_IMAGE_USAGE_SAMPLED_BIT;
-        ici.pNext = &export_metal_object_create_info;
-        VkImageObj mp_image_obj(m_device);
-        mp_image_obj.init(&ici);
-
-        metal_texture_info.bufferView = VK_NULL_HANDLE;
-        metal_texture_info.imageView = VK_NULL_HANDLE;
-        metal_texture_info.image = mp_image_obj.handle();
-        metal_texture_info.plane = VK_IMAGE_ASPECT_PLANE_2_BIT;
-        export_metal_objects_info.pNext = &metal_texture_info;
-        m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-VkExportMetalObjectsInfoEXT-pNext-06800");
-        vk::ExportMetalObjectsEXT(m_device->handle(), &export_metal_objects_info);
-        m_errorMonitor->VerifyFound();
-
-        if (ycbcr_conversion_extension) {
-            VkSamplerYcbcrConversionCreateInfo ycbcr_create_info = vku::InitStructHelper();
-            ycbcr_create_info.format = mp_format;
-            ycbcr_create_info.ycbcrModel = VK_SAMPLER_YCBCR_MODEL_CONVERSION_RGB_IDENTITY;
-            ycbcr_create_info.ycbcrRange = VK_SAMPLER_YCBCR_RANGE_ITU_FULL;
-            ycbcr_create_info.components = {VK_COMPONENT_SWIZZLE_IDENTITY, VK_COMPONENT_SWIZZLE_IDENTITY,
-                                            VK_COMPONENT_SWIZZLE_IDENTITY, VK_COMPONENT_SWIZZLE_IDENTITY};
-            ycbcr_create_info.xChromaOffset = VK_CHROMA_LOCATION_COSITED_EVEN;
-            ycbcr_create_info.yChromaOffset = VK_CHROMA_LOCATION_COSITED_EVEN;
-            ycbcr_create_info.chromaFilter = VK_FILTER_NEAREST;
-            ycbcr_create_info.forceExplicitReconstruction = false;
-
-            VkSamplerYcbcrConversion conversion;
-            err = vk::CreateSamplerYcbcrConversionKHR(m_device->device(), &ycbcr_create_info, nullptr, &conversion);
-            ASSERT_EQ(VK_SUCCESS, err);
-
-            VkSamplerYcbcrConversionInfo ycbcr_info = vku::InitStructHelper();
-            ycbcr_info.conversion = conversion;
-            ycbcr_info.pNext = &export_metal_object_create_info;
-            ivci.image = mp_image_obj.handle();
-            ivci.format = mp_format;
-            ivci.pNext = &ycbcr_info;
-            vkt::ImageView mp_image_view(*m_device, ivci);
-            metal_texture_info.image = VK_NULL_HANDLE;
-            metal_texture_info.imageView = mp_image_view.handle();
-            m_errorMonitor->SetDesiredFailureMsg(kErrorBit, "VUID-VkExportMetalObjectsInfoEXT-pNext-06802");
-            vk::ExportMetalObjectsEXT(m_device->handle(), &export_metal_objects_info);
-            m_errorMonitor->VerifyFound();
-            vk::DestroySamplerYcbcrConversionKHR(m_device->device(), conversion, nullptr);
-        }
-    }
-}
-#endif  // VK_USE_PLATFORM_METAL_EXT
 
 TEST_F(VkLayerTest, InvalidExtEnum) {
     TEST_DESCRIPTION("Use an enum from an extension that is not enabled.");
