+++ tests/unit/shader_spirv_positive.cpp
@@ -634,7 +634,7 @@ TEST_F(PositiveShaderSpirv, Spirv16Vulkan13) {
     VkShaderObj vs(this, kVertexMinimalGlsl, VK_SHADER_STAGE_VERTEX_BIT, SPV_ENV_VULKAN_1_3);
 }
 
-TEST_F(VkPositiveLayerTest, OpTypeArraySpecConstant) {
+TEST_F(PositiveShaderInterface, OpTypeArraySpecConstant) {
     TEST_DESCRIPTION("Make sure spec constants for a OpTypeArray doesn't assert");
     SetTargetApiVersion(VK_API_VERSION_1_1);
     RETURN_IF_SKIP(Init());
